+++ ray-curator/ray_curator/stages/download/text/common_crawl/commoncrawl.py
@@ -18,45 +18,10 @@ from typing import Literal
 
 from nemo_curator.datasets import DocumentDataset
 from nemo_curator.download.doc_builder import (
-    DocumentExtractor,
     download_and_extract,
 )
 from nemo_curator.utils.download_utils import get_common_crawl_urls
 from nemo_curator.utils.file_utils import expand_outdir_and_mkdir
-
-
-class CommonCrawlWARCExtractor(DocumentExtractor):
-    def __init__(
-        self,
-        algorithm: HTMLExtractorAlgorithm | None = None,
-        stop_lists: dict[str, frozenset[str]] | None = None,
-    ):
-        if algorithm is None:
-            algorithm = JusTextExtractor()
-
-        if stop_lists is not None:
-            self._stop_lists = stop_lists
-        else:
-            self._stop_lists = get_stop_list_dict()
-
-        self.algorithm = algorithm
-        super().__init__()
-
-    def extract(self, content: str) -> dict[str, str] | None:
-        html = decode_html(content)
-        if html is not None:
-            # Language detection and HTML extraction
-            lang = lang_detect(html)
-            text = None
-            if lang in self._stop_lists:
-                text = self.algorithm.extract_text(html, self._stop_lists[lang], lang)
-            if text is not None:
-                if len(text) > 0:
-                    text = "\n\n".join(text)
-                    return {"language": lang, "text": text}
-                else:
-                    return None
-        return None
 
 
 def download_common_crawl(  # noqa: PLR0913
